+++ tests/test_card_info_downloader.py
@@ -548,27 +548,42 @@ def test_set_wackiness_score(qtbot, card_db: CardDatabase, json_name: str, expec
     )
 
 
-def test_related_printings(qtbot, card_db: CardDatabase):
-    db = card_db.db
-    cards = [
+@pytest.mark.parametrize("cards, expected_pairs", [
+    ([
         "The_Underworld_Cookbook",
         "Food_Token",
         "Asmoranomardicadaistinaculdacar",
         "Bake_into_a_Pie",
         "Asmoranomardicadaistinaculdacar_2",
         "Food_Token_2",
-    ]
+     ], [
+        # The Food token (card id 2) is never a source, as that would pull all cards creating that token
+        (3, 1),  # Asmoranomardicadaistinaculdacar references The Underworld Cookbook by name
+        (1, 3),  # Back relation
+        (1, 2),  # Card mentions Food token
+        (4, 2),  # Card mentions Food token
+    ]),
+    ([
+         "Dungeon_of_the_Mad_Mage", "Zombie_Ogre", "Dungeon_Skeleton_Token",
+     ],[
+        (1, 3),  # The Dungeon itself can create a Skeleton Token
+        (2, 1),  # Zombie Ogre has Venture into the Dungeon
+        # Nothing else here:
+        # The Skeleton must not link to the Dungeon, and the Dungeon must not link to the Zombie Ogre
+    ]),
+])
+def test_related_printings(
+        qtbot, card_db: CardDatabase,
+        cards: typing.List[str], expected_pairs: typing.List[typing.Tuple[int, int]]):
+    db = card_db.db
+
     # Cards always relate to exact printings, but which one is chosen is rather arbitrary. E.g. The Underworld Cookbook
     # and Back into a Pie both create a Food token, but are set to different printings of that token card.
     fill_card_database_with_json_cards(qtbot, card_db, cards)
     assert_that(
         db.execute("SELECT card_id, related_id FROM RelatedPrintings").fetchall(),
         contains_inanyorder(
-            # The Food token (card id 2) is never a source, as that would pull all cards creating that token
-            (3, 1),  # Asmoranomardicadaistinaculdacar references The Underworld Cookbook by name
-            (1, 3),  # Back relation
-            (1, 2),  # Card mentions Food token
-            (4, 2),  # Card mentions Food token
+            *expected_pairs
         )
     )